Rename boilerplates to reflect syntax
