+++ js-cjs-then/.editorconfig
