+++ js-cjs-then/.eslintrc.json
