+++ js-esm-async-await/package.json
@@ -1,0 +1,34 @@
+{
+  "name": "js-esm-async-await-backend-api-boilerplate",
+  "version": "0.0.0",
+  "description": "Javascript (ES Module import/export syntax and async-await asynchronous code handing method) Backend API boilerplate template generated via Collabo Community's Create-collabo-app project",
+  "main": "src/server.js",
+  "scripts": {
+    "dev:local": "nodemon --exec babel-node src/server.js",
+    "dev:atlas": "nodemon --exec babel-node src/server.js",
+    "lint": "esw src --color",
+    "lint:watch": "npm run lint -- --watch"
+  },
+  "author": "Obiagba Mary",
+  "license": "ISC",
+  "dependencies": {
+    "cors": "^2.8.5",
+    "dotenv": "^16.0.3",
+    "esm": "^3.2.25",
+    "express": "^4.18.2",
+    "mongoose": "^7.0.5",
+    "morgan": "^1.10.0"
+  },
+  "devDependencies": {
+    "@babel/cli": "^7.21.0",
+    "@babel/core": "^7.21.4",
+    "@babel/eslint-parser": "^7.21.3",
+    "@babel/node": "^7.20.7",
+    "@babel/preset-env": "^7.21.4",
+    "chalk": "^4.1.0",
+    "eslint": "^8.39.0",
+    "eslint-plugin-import": "^2.27.5",
+    "eslint-watch": "^8.0.0",
+    "nodemon": "^2.0.22"
+  }
+}
