+++ ts-esm-async-await/package.json
@@ -1,8 +1,8 @@
 {
-  "name": "node-mongo-api-boilerplate-template-ts",
-  "version": "1.0.0",
-  "description": "@code-collabo node-mongo API boilerplate template - typescript",
-  "main": "src/server.[atlas/local].ts",
+  "name": "ts-esm-async-await-backend-api-boilerplate",
+  "version": "0.0.0",
+  "description": "Typescript (ES Module import/export syntax and async-await asynchronous code handing method) Backend API boilerplate template generated via Collabo Community's Create-collabo-app project",
+  "main": "src/server.ts",
   "scripts": {
     "dev:local": "nodemon --exec ts-node src/server.ts",
     "dev:atlas": "nodemon --exec ts-node src/server.ts",
