+++ ts-esm-async-await/src/lib/errors/CustomError.ts
