+++ ts-esm-async-await/src/server.ts
@@ -1,6 +1,6 @@
 import mongooseConnect from './db.connect';
 import { app as app } from './app';
-import { connectionType } from '../node-mongo-helpers';
+import { connectionType } from './node-mongo-helpers';
 
 
 const port = connectionType().port || 3000;
